requirements.txt
imutils==0.5.3
opencv_python==4.1.1.26
requests==2.21.0
SpeechRecognition==3.8.1
tkcalendar==1.6.1
gTTS==2.1.1
beautifulsoup4==4.9.0
Pillow==7.1.2
